Add extension points in the SDK
